// src/buildfx/dotnet/enums.ts
export enum VisualStudioVersion {
    Latest = "latest",
    VisualStudio2022 = "17.0",
    VisualStudio2019 = "16.0",
    VisualStudio2017 = "15.0",
    VisualStudio2015 = "14.0",
}

export enum VisualStudioEditions {
    All = "*",
    Community = "Community",
    Professional = "Professional",
    Enterprise = "Enterprise",
    BuildTools = "BuildTools",
}